==> 12-prj-supervised-regression/02-feature-engineering.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import sklearn

The importing data should be generated from the previous lecture(.ipynb file)


# Import training and validation sets (sorted)
df_tmp = pd.read_csv("../data/bluebook-for-bulldozers/TrainAndValid-after-01.csv",
                 low_memory=False,
                 parse_dates=["saledate"])

### Add datetime parameters for `saledate` column

df_tmp["saleYear"] = df_tmp.saledate.dt.year
df_tmp["saleMonth"] = df_tmp.saledate.dt.month
df_tmp["saleDay"] = df_tmp.saledate.dt.day
df_tmp["saleDayOfWeek"] = df_tmp.saledate.dt.dayofweek
df_tmp["saleDayOfYear"] = df_tmp.saledate.dt.dayofyear

df_tmp.head().T

# Now we've enriched our DataFrame with date time features, we can remove 'saledate'
df_tmp.drop("saledate", axis=1, inplace=True)

# Check the values of different columns
df_tmp.state.value_counts()

df_tmp.head()

len(df_tmp)

## 5. Modelling 

We've done enough EDA (we could always do more) but let's start to do some model-driven EDA.

# Let's build a machine learning model
from sklearn.ensemble import RandomForestRegressor

model = RandomForestRegressor(n_jobs=-1,    #  -1 means using all processors
                              random_state=42)

X = df_tmp.drop("SalePrice", axis=1)
y = df_tmp["SalePrice"]

# model.fit(X, y)   # Error

# Export current tmp DataFrame
df_tmp.to_csv("../data/bluebook-for-bulldozers/TrainAndValid-after-02.csv", index=False)
