==> 12-prj-supervised-regression/06-process-predict.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import sklearn

# Import training and validation sets (modified)
df_tmp = pd.read_csv("../data/bluebook-for-bulldozers/TrainAndValid-after-04.csv",
                 low_memory=False)

# Split data into training and validation
df_val = df_tmp[df_tmp.saleYear == 2012]
df_train = df_tmp[df_tmp.saleYear != 2012]

len(df_val), len(df_train)

# Split data into X & y
X_train, y_train = df_train.drop("SalePrice", axis=1), df_train["SalePrice"]
X_valid, y_valid = df_val.drop("SalePrice", axis=1), df_val["SalePrice"]

X_train.shape, y_train.shape, X_valid.shape, y_valid.shape

import pickle

# Load a saved model
ideal_model = pickle.load(open("05_ideal_random_forest_regressor_model.pkl", "rb"))

## Make predictions on test data

# Import the test data
df_test = pd.read_csv("../data/bluebook-for-bulldozers/Test.csv",
                      low_memory=False,
                      parse_dates=["saledate"])

df_test.head()

# # Make predictions on the test dataset
# test_preds = ideal_model.predict(df_test)

# # ValueError: could not convert string to float: 'Low'

df_test.isna().sum()

df_test.info()

df_test.columns

X_train.columns

### Preprocessing the data (getting the test dataset in the same format as our training dataset)

def preprocess_data(df):
    """
    Performs transformations on df and returns transformed df.
    """
    df["saleYear"] = df.saledate.dt.year
    df["saleMonth"] = df.saledate.dt.month
    df["saleDay"] = df.saledate.dt.day
    df["saleDayOfWeek"] = df.saledate.dt.dayofweek
    df["saleDayOfYear"] = df.saledate.dt.dayofyear
    
    df.drop("saledate", axis=1, inplace=True)
    
    # Fill the numeric rows with median
    for label, content in df.items():
        if pd.api.types.is_numeric_dtype(content):
            if pd.isnull(content).sum():
                # Add a binary column which tells us if the data was missing or not
                df[label+"_is_missing"] = pd.isnull(content)
                # Fill missing numeric values with median
                df[label] = content.fillna(content.median())
    
        # Filled categorical missing data and turn categories into numbers
        if not pd.api.types.is_numeric_dtype(content):
            df[label+"_is_missing"] = pd.isnull(content)
            # We add +1 to the category code because pandas encodes missing categories as -1
            df[label] = pd.Categorical(content).codes+1
    
    return df

# Process the test data 
df_test = preprocess_data(df_test)
df_test.head()

# # Make predictions on updated test data
# test_preds = ideal_model.predict(df_test)

# # ValueError: X has 101 features, but RandomForestRegressor is expecting 102 features as input.

X_train.head()

# We can find how the columns differ using sets
set(X_train.columns) - set(df_test.columns)

# Manually adjust df_test to have auctioneerID_is_missing column
df_test["auctioneerID_is_missing"] = False
df_test.head()


# df_test_tmp = pd.DataFrame(data=None, columns=X_train.columns, index=None)

# for col in X_train.columns:
#     df_test_tmp[col] = df_test[col]

# df_test = df_test_tmp
# # df_test_tmp.compare(df_test)
# # df_test_tmp.head().T

# df_test.head().T

# Make predictions on the test data
test_preds = ideal_model.predict(df_test)

test_preds

test_preds

We've made some predictions but they're not in the same format Kaggle is asking for: https://www.kaggle.com/c/bluebook-for-bulldozers/overview/evaluation

# Format predictions into the same format Kaggle is after
df_preds = pd.DataFrame()
df_preds["SalesID"] = df_test["SalesID"]
df_preds["SalesPrice"] = test_preds
df_preds

# Export prediction data
df_preds.to_csv("../data/bluebook-for-bulldozers/test_predictions-after-06.csv", index=False)
